services/product_service/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
mysqlclient==2.1.1
django-cors-headers==4.1.0
PyJWT==2.6.0
pika==1.3.2
requests==2.28.2
gunicorn==20.1.0
Pillow>=9.0.0